reporter: Improve flaky job Sentry reporting
This patch has the following changes in order to accommodate better
visibility in the flaky jobs. First and foremost, the reporting is now
on a *per flaky job basis* (instead of a generic "Flaky jobs detected")
exception.

In this way, every event for a specific test is grouped with its
predecessors thus enabling chronological sorting, and volume of events
per test case.

A tag "flaky" was also added in the event so that the flaky job exceptions
can be filtered out better.