queue: Enable fast failure
These commits adds support on the Queue class to terminate builds after
a certain threshold of failures is surpassed. This configuration is
controlled via the `Queue#publish` method and it is persisted on Redis
Hash[1]. The persistence is added so that the value is shared between
different components that use the same queue instance. (e.g. worker and
reporter)

The feature can be enabled via the `--fail-fast` flag which is passed
onto the worker. This is in par with RSpec's `--fail-fast`[2] which will
terminate the execution of the suite after certain failed examples.
This is an optional flag, and will be nil if not set.

[1]: https://redis.io/topics/data-types#hashes
[2]: https://relishapp.com/rspec/rspec-core/docs/command-line/fail-fast-option