Display more information regarding flaky tests
Flaky tests may fail, but the rspec output is lost in the process.

This commit enriches the sentry event with the spec seed, expectation,
backtrace, as well as with the id of the worker in which the spec failed.

The devs can see what was expected but most importantly what was produced.

This is a temporary solution until we introduce a better way to capture the
distributed worker outputs and provide a true and easy way to reproduce the
flaky tests.

We drop the rerun commands entirely since they are not successful in
correctly reproducing the errors are I originally thought. We go back to a
simpler reporter output.