Help users reproduce flaky tests
Originally, a flaky test would appear in the report by its just example id
which is not helpful at all.

Then we merged 9f6816e which tried to replace the flaky example id with a
reproduction command. The command did not work however because the premise
was wrong. You see, to correctly reproduce a flaky test you need more than
the seed and the test file. You also need all the files the preceded the
flaky test, up to the point in time when you reset rspec and the databases.
For rspecq, this point in time is the start of a worker.

Both of the above issues are mitigiated:
1. Flaky tests are reported by their example location, which by itself does
not offer anything, but at least it shows the developer which test exactly
is flaky.
2. Provides a reproduction command for each flaky test, which is the gist.

The reproduction (or rerun) command, consists of the seed and the files originally tested by the worker when the test failed.