Introduce 'reproduction' argument
The new argument modifies how the rspecq worker publishes the files and
examples to the queue.

By detault it first merges individual examples into files and then
depending on whether the `update-timings` is passed, it will either
publish based on the timings or randomly shuffly them.

If the `reproduction` argument is passed, the worker will ignore timings and
randomness and will publish the files exactly as given.

This feature is very important in order for a developer to be able to
successfully reproduce a flaky test as reported by the reporter in a previous
run.