Allow users to have shorter flaky report
The rerun command for the flaky tests can be pretty big. This becomes
a bigger problem when there are more flaky tests appearing.

We allow users to skip the rerun command from the summry report by
setting the env variable `RSPECQ_REPORTER_RERUN_COMMAND_SKIP`. The location
and worker are still repoted.

The sentry payload remains the same.