Exit immediately on SIGTERM instead of gracefully!
It seems that ECS/Fargate will send SIGTERM on ALL containers which
will not wait for the job to finish before exiting.

For example mysql will stop during the gracefull shutdown which
invalidates or blocks the spec file. Other workers will have to
wait for the heartbeat timeout to elapse before they pick it up.

With this commit we do not gracefully wait for the job to finish,
instead we immediately requeue the job and remove the worker from
the queue state ("running", "heartbeats") and leave other workers
to re-run it.