reporter: Report unique events to Sentry per spec
Sentry messages are displayed as issues in the issues stream,
with the message serving as the issue's title.
Up until now, the name of the file containing the failing spec was
included in the event message, thus all failures for each spec file were
grouped together under the same event.

With this commit, a SHA1 digest is created for each spec (spec filename
+ spec name) and included in the event message, leading to unique
Sentry events per spec file.