mark build status in redis
Now, the build status is not stored anywhere, the reporter decides
it by inspecting the state of the queue. Having the final state stored
in redis, decouples the build from the reporter, and allows us to
query redis and get the build status.

With the patch the status (failed or success) is stored by the first
worker that finds out that the build is over.

This status is updated by a lua script to ensure atomicity. The script
itself is defensive, and rechecks that the build is in fact over. This
is not strictly necessary because try_mark_finished is called after the
worker has checked that.